User model and user authentication moved to App\User namespace
