+++ _api_app/app/User/UserModel.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App;
+namespace App\User;
 
 use Illuminate\Auth\Authenticatable;
 use Laravel\Lumen\Auth\Authorizable;
 use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
 use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableContract;
 
-class User implements
+class UserModel implements
     AuthenticatableContract,
     AuthorizableContract
 {
