+++ _api_app/bootstrap/app.php
@@ -79,7 +79,7 @@ $app->routeMiddleware([
 */
 
 // $app->register(App\Providers\AppServiceProvider::class);
-$app->register(App\Providers\AuthServiceProvider::class);
+$app->register(App\User\UserAuthServiceProvider::class);
 // $app->register(App\Providers\EventServiceProvider::class);
 
 /*
